registerをリファクターした
