+++ app/src/main/java/com/kyotob/client/constants.kt
@@ -7,3 +7,6 @@ const val USER_DATA_KEY = "userData"
 const val USER_NAME_KEY = "userName"
 const val USER_SCREEN_NAME_KEY = "screenName"
 const val TOKEN_KEY = "accessToken"
+
+const val IMAGE_PREFERENCE_KEY = "image"
+const val IMAGE_PATH_KEY = "imagePath"
